add specu_decoding support for Mixtral,qwen2,qwen2-moe
